:rocket: Add endpoint to log users through facebook

+++ src/main/java/com/watchlist/backend/model/Role.java
@@ -8,8 +8,8 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "role")
 public class Role {
-    public static final int ADMIN = 1;
-    public static final int WATCHER = 2;
+    public static final long ADMIN = 1;
+    public static final long WATCHER = 2;
 
     @Id
     private long id;
